+++ apps/api/controllers/data/search.rb
@@ -5,7 +5,15 @@ module Api::Controllers::Data
     expose :blobs
 
     def call(params)
-      @blobs = BlobRepository.new.all
+      repository = BlobRepository.new
+      @blobs = []
+      if params[:format]
+        @blobs = repository.by_format(params[:format])
+      elsif params[:tags]
+        @blobs = repository.by_tags(params[:tags])
+      else
+        @blobs = repository.all
+      end
     end
   end
 end